Fixes #11
