+++ index.html
@@ -15,7 +15,7 @@
     <div id="menu">
       <img src="" id="delete">
       <ul id="list-select">
-        <li id="list-default"></li>
+        <li id="list-default">Hej</li>
       </ul>
     </div>
   </body>
